Classify_object/requirements.txt
tensorflow
fastapi
uvicorn[standard]
pillow
python-multipart
